+++ tests/testthat/test-metrics.R
@@ -1,8 +1,8 @@
 test_that("network metrics work", {
-  sim_data <- simulate_edge_model("binary", aggregated = TRUE)
+  sim_data <- simulate_bison_model("binary", aggregated = TRUE)
   df <- sim_data$df_sim
 
-  fit_edge <- edge_model(
+  fit_edge <- bison_model(
     (event | duration) ~ dyad(node_1_id, node_2_id),
     data=df,
     data_type="binary_conjugate",
@@ -10,63 +10,53 @@ test_that("network metrics work", {
   )
 
   x <- expect_warning(
-    draw_node_metric_samples(fit_edge, "strength"),
+    extract_metric(fit_edge, "node_strength"),
     regexp=NA
   )
 
-  fit_edge <- edge_model(
+  fit_edge <- bison_model(
     (event | duration) ~ dyad(node_1_id, node_2_id),
     data=df,
     data_type="count_conjugate",
-    priors=get_default_priors("binary_conjugate")
+    priors=get_default_priors("count_conjugate")
   )
 
   x <- expect_error(
-    draw_node_metric_samples(fit_edge, "not_a_metric", standardise=TRUE),
-    regexp="Network metric does not exist"
+    extract_metric(fit_edge, "not_a_metric", standardise=TRUE)
   )
 
   x <- expect_warning(
-    draw_node_metric_samples(fit_edge, "strength", standardise=TRUE),
+    extract_metric(fit_edge, "node_strength", standardise=TRUE),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_node_metric_samples(fit_edge, "betweenness"),
+    extract_metric(fit_edge, "node_betweenness"),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_node_metric_samples(fit_edge, "eigenvector"),
+    extract_metric(fit_edge, "node_eigen"),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_node_metric_samples(fit_edge, "degree[0.5]"),
+    extract_metric(fit_edge, "node_degree[0.5]"),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_network_metric_samples(fit_edge, "social_differentiation"),
+    extract_metric(fit_edge, "global_cv"),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_network_metric_samples(fit_edge, "weighted_density"),
+    extract_metric(fit_edge, "global_density"),
     regexp=NA
   )
 
   x <- expect_warning(
-    draw_network_metric_samples(fit_edge, "standard_deviation"),
-    regexp=NA
-  )
-
-  expect_output(
-    print(x)
-  )
-
-  expect_warning(
-    plot(x),
+    extract_metric(fit_edge, "global_std"),
     regexp=NA
   )
 })